refactor: move files from utils/ to types/
